more lint happiness
git-svn-id: file:///svn/unbound/branches/edns-subnet@2818 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ daemon/unbound.c
@@ -521,7 +521,7 @@ perform_setup(struct daemon* daemon, struct config_file* cfg, int debug_mode,
 		 * also resource limits from login config, but we
 		 * still call setresuid, setresgid to be sure to set all uid*/
 		if(setusercontext(NULL, pwd, uid,
-			LOGIN_SETALL & ~LOGIN_SETUSER & ~LOGIN_SETGROUP) != 0)
+			(unsigned int)LOGIN_SETALL & ~LOGIN_SETUSER & ~LOGIN_SETGROUP) != 0)
 			log_warn("unable to setusercontext %s: %s",
 				cfg->username, strerror(errno));
 #endif /* HAVE_SETUSERCONTEXT */
